Patch core's create-sibling-ghlike to use new credential system
The switch makes the storage of a newly entered credential
conditional on a successful authorization, in the spirit of
datalad/datalad#3126.

Moreover, stored credentials now contain a `realm` property that
identified the API endpoint. This makes it possible to identify
candidates of suitable credentials without having to specific
their name, similar to a request context url used by the old
providers setup.

This automatic realm-based credential lookup is now also implemented.
When no credential name is specified, the most recently used
credential matching the API realm will be used automatically.
If determined like this, it will be tested for successfull
authorization, and will then be stored again with an updated
'last-used' timestamp.