add a version of create-sibling-webdav
Create a set of two remotes, one for git annex export and one for
datalad push. One is called <name>-vcs and is meant to contain the
git-repository info. The other one is called <name>-tree and is
meant to contain the annex-exported tree.

Credential manager is still
"datalad.downloaders.credentials.UserPassword".