start socket listening before exiting __enter__
If socket binding and listening are performed
inside the thread that is started in '__enter__'
of the 'WebDAVPath' context manager, the main
thread might try to connect to a socket in a
'weird' state. That might lead to a hanging
client.

This commit ensures that binding and listening
are performed before the server thread is
started and before we return to the main thread.