Implement --existing switch
Implements `--existing` in order to skip or reconfigure already existing
git or annex special remotes as determined by their name. Skip would not
skip the entire dataset in such a case but only the creation of the
sibling that's actually in conflict.

This now also considers not yet enabled special remotes with the same
name as 'existing' ones.