Patch -core to be able to auto-deploy credentials for enabling special remotes
Two patches are needed.

- a patch of `siblings.py:_enable_remote` simple removes the
  webdav-specific credential handling in core
- a patch of `AnnexRepo.enable_remote` enables query and auto-deploy
  of new-style credentials

The second patch is ugly, because of https://github.com/datalad/datalad/issues/5162

However, the first patch removes the only cause for this issue, so
theoretically, the patch could even remove the env parameter. But
leaving for now.

Manual credential entry on `enableremote` is not implemented here, but
easily possible following the patterns from `datalad-annex::` and
`create_sibling_webdav()`