add tests for '_get_url_credential'
This commit adds tests for the logic in
'_get_url_credential'.

It tests that credentials are derived
from WEBDAV-URLs if the credential name
is not provided.

It checks that the credential name is
used to derive credentials if it is provided.

It checks that credentials are
derived from other properties, e.g. realm,
if neither a name is provided, nor are there
credentials associated with the URL