BF: Fix --recursive
ATM, `foreach_dataset` would always go into the first level of datasets
regardless of --recursive, due to its internal call to `subdatasets`.
Account for that, by extracting non-recursive case, but keep the
original call for the recursive case. Not a datalad_next specific
helper, since this should be addressed by `foreach_dataset` in core and
didn't want to decouple from that.