handle main-case in top-level execution path
This commit implements a suggestion from Michael Hanke,
i.e. to handle the main-case of a function in the top
level exection path.

The commit modifies '_transfer_data' to handle the
"non-interesting" standard push in a leading "if not"-
clause, which contains a return statement.

The main-case, i.e. export, is then handled in the
remainder of the function.