BF: Test operated on PWD of test execution
This was a faulty pattern, making `create-sibling-webdav` rely on PWD to
figure out a dataset to run on. While this is fine from within the
command, tests need to make sure to provide the correct env etc. to run
from within.
Otherwise this pattern interferes with repos of the user running the
tests.