Unify result action for git-remotes
This replaces the `add-sibling` and `configure-sibling` actions coming
from the siblings command used for git-remotes. And adjusts the results
for special remotes accordingly.

Now the action is either `create_sibling_webdav` or
`reconfigure_sibling_webdav` with an `.storage` suffix in case of special
remotes. In opposition to before, this unifies the action field across
statuses.