Support annex private repositories
This patch enables datalad support for annex private repositories (or
rather the creation of such via `datalad-create`, `datalad-clone` and
`datalad-get`. For such repositories neither availability nor the
existence of this location will be recorded in the git-annex branch.

This implementation introduces a datalad config
`datalad.annex.init.private` (default: false), that is considered
whenever we annex-init a repository so that the respective annex config
`annex.private` is set before the call to `git annex init`. The setting
is therefore respected by `create`, `clone` and `get`.
In addition, if the datalad config was true, it is recorded in the local
`.git/config`. That way subsequent `clone`/`get` calls for subdatasets
to inherit that setting by default, just like they would consider the
value of the `reckless` parameter that was used with the superdataset.