Use blunt force to drop keys on windows
This is a rather draconian workaround: Instead of dropping the repoannex
keys, it wipes out the entire object tree of the local repoannex utility
repo.

This addresses
https://git-annex.branchable.com/bugs/Fails_to_drop_key_on_windows___40__Access_denied__41__/?updated
which is still happing with today's git-annex snapshot.

This approach possible, because this utility repo is constructed on the
fly from the URL specification, and only contains the two keys that need
to be dropped. However, given the bluntness of the approach, it is
limited to the platform on which the original issue is happening:
windows.

Closes datalad/datalad-next#26