Fix results when pushing to export remote
The push-to-export-remote patch of core's push command failed to handle
the results from `AnnexRepo._call_annex_records_items_` correctly.
It was not accounting for the fact, that this are JSON records as
reported by git-annex (thus ignoring a bunch of fields), which should be
done by `datalad.interface.results.annexjson2result`.
Furthermore it was unconditionally setting the status to `ok` expecting
a `CommandError` otherwise. This is not how the generator works in the
general case, though. It does yield `error` records and proceeds if it
can. `CommandError` still can happen but it's not the only thing leading
to an error result.

This led to `push` always "succeeding" with an `ok` result, rendering
all tests of export remotes useless, that rely on `push` failing if
there was something wrong.

The respective tests made were based on the same, wrong assumptions
about `_call_annex_records_items_`.

This commit (hopefully) fixes this.
(Closes #88)