Add an example of adding a user_password credential
This adds an example of how to set a new credential of type
user_password, with a given user property, to the examples if the
credentials command.

It shows that type and user need to be defined as properties, and
entry of the secret is interactive. While it was probably clear from
the other examples and command description, now it's explicit.