Don't split prompt and question in CredentialManager
In opposition to the `Credential`'s in datalad core, the
CredentialManager splits the prompting text and the input retrieval when
asking for a property or secret by calling `ui.message` and then
`ui.question`. While this is kinda fine with CLI, other `ui` backends
like gooey can't handle this properly (leading to
https://github.com/datalad/datalad-gooey/issues/340).

The `title` parameter of `ui.question` is intended for the prompting
text, which is somewhat misleading by name.