Extend `Constraint` with `for_dataset()` context switcher
It generates a new constraint from another one that is tailured
to a particular `Dataset` instance. By default it return `self`,
ie. the same constraint.

This can be useful to tighten validation, once a context is (more)
known -- while maintaining semantics.

For example, an `EnsureSiblingName` is equal to an
`EnsureStr(min_len=1)` without a dataset context, but can be specialized
to `EnsureChoice()` given a particular dataset.

The output of `for_dataset()` would generally not be back-convertible.
A typical usage pattern would be to keep the original `Constraint`
instance around to be able to re-tailor it to any `Dataset` instance
when they become available.