Fix test fixture appliance
The order of execution of the former markup for the base class vs
derived class somehow changed, leading to the base class' injection
overwriting what the subclass already set for its `path`.
Unsure whether that is depend on python or pytest version.

In any case, that means a setup independent of that order is needed.
Hence, instead of overwriting the `path` class attribute, have a fixture
that dynamically decides based on a another, constant class attribute
defining which kind of tree the test case wants.