Parameter validation in datalad.interface.utils._execute_command_()
It inspects an `Interface` class whether it has a `validate_args()`
method (for now), and if so, feeds all parameters through it.

It also changes the actual command execution by passing all parameters
as keyword-arguements (only filtering those that were present in the
`exec_kwargs` dict (generic result handling parameters).

All this is subject to further changes and improvements.