New `EnsureDataset` constraint
Unlike the one from datalad-core it actually returns `Dataset`
instances.

However, this comes with problems that arise from the need to patch
`resolve_path()` in order to maintain the semantics of relative
path arguments establish in DataLad.

With this constraint, a command always receives a Dataset instance.
An internal call to `resolve_path()` cannot decide (based on the type)
which resolution strategy to use. Hence the Dataset instance itself
needs to carry this information.

That, in turn, is problematic, because a Dataset instance is shared
across all contexts based on a unique-path criterion.

See docstring of `EnsureDataset` for more details.