Let `credentials()` and `CredentialManager.set()` prompt for name
This change enables users and commands to leave a credential name
unspecified. If possible (interactive session), a user will be prompted
for a name (until a unique name was entered), or "skip" was indicated.

If entering a credential name is not possible, a `ValueError` is raised.

`CredentialManager.set()` is enhanced with an additional `_context`
argument that is included in a name prompt to provide context to a user
what kind of credential is being saved. Calling commands typically have
a better idea on that vs barfing a potentially complex credential with
all properties in the terminal.

A new `_suggested_name` argument for `CredentialManager.set()` can
be used by callers to suggest a default name. If the suggestion conflicts
with an existing credential it is ignored.

Closes datalad/datalad-next#130