Add file:// URL support to `download_file()`
Feature parity with http:// URLs:

- simultaenous hashing
- progress reporting

The implementation is based on `shutil.copyfileobj()`, so it may not
achieve the most optimal performance on each platform. However,
it does use `shutil`'s wisdom re determining the size of the chunks
to read/write at once.

It is written to be compatible with file:// URLs as produced by
`Path.as_uri()`.

Ping datalad/datalad-next#125