Switch download-file mapping spec to space-delimited
tab-delimited cannot be easily typed into a command-line. It was
possible via JSON-lines encoding or piping via `stdin`. However,
"download-a-single-url-to-a-custom-path" is such a standard use
case that it feels not-right to make it this complicated.

The best compromise (as presented here), seems to be space-separation,
and thereby allow URL-path pairs to be specified via a single quoted
string.