add timeout and error handling to ssh upload
This commit adds timeout handling to ssh upload.
That means, it extends the protocol that is used
with a timeout method and adds a timeout keyword
argument to _SshCat.run() that is forwarded to
the runner.

In addition, this commit add code to exhaust the
generator returned by _SshCat.run(), in order to
trigger CommnandError exceptions if the subprocess
returns a non-zero exit status. This can be done
by simply using `tuple(ssh_runner_generator)`
because we do not expect any output, just a
StopIteration when the subprocess exits.

This commit does not yet expose timeout at the
`UrlOperations`-interface. This will be done
in one of the following commits