add support for ssh upload from stdin
This commit adds support for stdin as upload
source when uploading to a ssh URL. A simple
test for uploading from `stdin` is added as
well.

The code assumes that
`datalad_next.url_operations._progress_report_start`
is capable of handling an `expected_size` of `None`,
and that this triggers a report for an upload of
unknown size.