fix a bug where ssh subprocesses steal stdin
This commit fixes a bug that was triggered when
an annexremote process used ssh to communicate
with a remote.

The issue was the the annexremote process would
not receive all data that was sent over stdin.
The reason for that was that the ssh-subprocesses
inherit stdin from the parant process, i.e. from
annexremote, and suck in data before the annexremote
can read it. This is due to a `None`-argument in the
subprocess.Popen stdin keyword argument. That leads
to the subprocess inheriting stdin from the parent
process. Instead `subprocess.DEVNULL`, would be a
proper argument.

This should also be done in the `ThreadedRunner`-
code in datalad (which would make this commit
unnecessary).