Constraint for joint validation of a full command parameterization
This provides a solution to the `EnsureMappings` proposal from
https://github.com/datalad/datalad-gooey/issues/311, albeit in
a command-specific fashion.

This now establishes an interface to perform additional validation steps
(uniformly across all APIs) after per-parameter validation was completeded
successfully.

The underlying validator class could also be a source for documentation
to be exposed to users via the help-generation tooling, but this is not
done here, because they largely sit in datalad-core.

The utility of this approach is demo'ed on the `credentials` command,
which moves information from one parameter to another, depending
on the origin API. This is fully captured, and no validation or
type conversiion code remains in the actual command implementation
(`__call__`).

The impact on implementation complexity for commands that do not
require validation of inter-parameter dependencies is demo'ed
for the `download` command.

Ping #115