Importless declaration of URL handler
Previously, the dict based on which `AnyUrlOperations` selects a
handler must contain a handler class as values. This made it necessary
to import everything first, and thereby creates an avoidable startup
cost.

Now one declares a tuple with a module name and a class name to
import on demand as values.

This avoids unconditional dependencies on special interest packages.

Closes datalad/datalad-next#175