Support arbitrary FSSPEC customizations for S3 access
This includes, but is not limited to, being able to specify
whether or not anonymous access should be attempted (first).

This change paves the way for endpoint customizations and
anything else that FSSPEC exposes.

Moreover, when an explicit `credential` identifier is given,
the boto-based attempt to locate credentials or try
anonymous access first is skipped, and a credential is looked
up and provisioned immediately.