Introduce `FsspecUrlOperations(block_size=...)` parameter
See docs inside. At present it is unclear how relevant this will be
in practice. However, different filesystem caching settings
impact performance substantially (empricial observation). If caching
is turned off entirely, this parameter is the only way to specify
chunk sizes (for reading). Hence I consider it a useful thing to have
in general -- and it is cheap.

This change also flips the default download method from iteration
over a file pointer to reading chunks of a specific size. Ths has been
found to be more performant in some cases.