Draft of loading URL handlers from configuration
Here is an example:

```cfg
[datalad "url-handler.(^|.*::)s3://mybucket/"]
    class = datalad_next.url_operations.fsspec.FsspecUrlOperations
    kwargs = {\"fs_kwargs\": {\"s3\": {\"anon\": false}}}
```

This defines a dedicated handler for accessing the AWS S3 bucket
'mybucket', and turns off anonymous access for this bucket specifically.

A regex-match expression is directly included in the config subsection
name (following the pattern of the standard Git `url.<base>.insteadOf`
configuration pattern).

The only mandatory configuration property is `class`. It points to an
importable class implementing the `UrlOperations` API.

If necessary, arbitrary keyword-arguments for the class constructur
can be specified, JSON-encoded, in a `kwargs` configuration item.

On instantiating the `AnyUrlOperations` handler, the configuration is
parse using the `ConfigManager` given to the handler, and configuration
based handler specifications overwrite and extend any handlers declare
in the built-in handler registry.

All active handlers are logged at level 8. Invalid handler
configuration is ignore and reported at 'debug' log level.

Closes datalad/datalad-next#217