`archivist` git-annex special remote
This is a replacement for the implementation of the `datalad-archives`
remote. In addition to its predecessor, it reduces the storage overhead
from 200% to 100% by doing partial extraction from fully downloaded
archives.

Ultimately, it will support sparse/partial access to remote archives,
avoiding any storage overhead, and the requirement to unconditionally
download full archives (see #215).

This implementation is trying to be efficient by:

- trying to fulfill requests using locally present archives
- trying to download smaller archives before larger ones

This implementation is aiming to be extensible by:

- using `ArchiveOperations` as a framework to uniformly implement
  (partial) access to local (or remote) archives.

Support for a number of corner cases is implemented (e.g., registered
availability from an archive that actually does not contain a given
key), but there are presently no tests for this, yet.