`datalad.archivist.legacymode` config switch
When this is set, the archivist special remote will no only do
what the datalad-archives special remote used to do, but will
actually run that particular implementation.

This is generally not desirable. It will disable all additional
features that the archivist would provide. However, in scenarios
where the nitty-gritty details of the previous implementation
matters, this can still be useful.

The implementation of this legacy switch in here should be cheap
enough to provide this fallback option without any significant costs.

The import of the legacy code is also localized.