Simplified configurability implementation, change item name
Remove the ability to support different configurations for more than
one configured `archivist` special remote in a single repository.

While this is technically possible, I could not come up with an actual
use case that would justify adding the complexity (even if it is not
much for this particular case).

My issue is that having this one item be configurable, it would dictate
that all others would also be. Until a real use case is found, this cost
can be avoided.

Given that the original `datalad-archives` remote is around since
forever, and no demand for this was ever expressed (as far as I can
see), it is unlikely that there is a need for this.