Allow for opt-in parameter default validation
This change addresses a key shortcoming of the new command parameter
validation approach. Previously, it did not process a parameter's
default value (at all), and required each command implementation to
handle a default value directly.

This caused an inconvenience for the ubiquitous `dataset` parameter,
which often defaults to `None` to indicate "will discover in CWD".
However, given that `None` is the default, this discovery would need
to be implemented in each command, rather than using the standard
`EnsureDataset`.

With this change, `EnsureCommandParameterization` can be
parameterized with a new `validate_defaults` argument to cause
processing of an individual parameter's default.

Here, this is implemented for the `create_sibling_webdav()` command
as a demonstration. Likely other commands need a similar change.

Closes #225