Joint parameter validation
This extends and formalizes the validation of parameter sets (beyond
independent validation of individual parameters). The most significant
change here is the introduction of `ParameterConstraintContext`,
a dataclass representing such a validation context (see docstring).

Based on this new class, all other plain-text parameter label
identifiers have also been replaced with this class. This makes
reporting more uniform, and also makes the `__all__` context label
obsolete.

A basic test to demo this is included.