Change on_error mode names for parameter validation
This is an anticipation of a future implementation change,
where mutually independent checks are executed in parallel.

In such a case, 'immediately' does not make any sense anymore.
Likewise, 'raise-before-joint' is not necessary, because
via the `ParameterConstraintContext` specifications we know
exactly which validations depend on which parameters, and
we can reliably determine which ones can still make sense
to run (or not, because an earlier validation step for a
parameter has already failed, and we cannot assume that the
input assumptions of a later check are valid).

This leaves: 'raise-early', 'raise-at-end'