Only work for complete credentials in `DataladAuth`
Previously, the last used credential matching a `realm` was used
unconditonally. With this change all matching credentials are filtered
to exclude any credentials without a secret set.

A secret might not be present for any reason (removed by an external
process, or never set to begin with). Regardless of the cause for such
an anomaly, such a credential can never work for authentication.
Therefore, we exclude them immediately. Should no other credential
matchin a realm provide a secret, prompting for credential entry is
attempted.

Closes #245