Make reporting behavior match documentation
Only report on credentials that have a secret. Issue only a debug
message on credential fragments without one.

The recorded use case for the previous behavior was the discovery of
credential fragments for clean-up. However, that was nohow properly
implemented, and it seems cleaner to ignore random cruft than have it
impose handling complexity on a CredentialManager user.

Now `CredentialManager.get()` reports a credential that has a secret,
or it does not report at all (i.e. None).