Make dataset-look behavior of `credentials` identical to `download`
When called with `dataset=None`, `credentials()` would auto-locate
a dataset containing CWD (and use its configuration), whereas
`download()` would use the global configuration in this case.

This change makes the behavior uniform and chooses what `download()`
did as the model behavior.

This uniformatization is meaningful on its own.

However, due to https://github.com/datalad/datalad/issues/7299 it is
also a necessity, because otherwise credentials deposited by
`download()` remain invisible to `credentials()` -- which is the
superficial trigger for #249

Closes #249