Test fixture to verify an unmodified global git config
Along the lines of the plan outlined in
https://github.com/datalad/datalad/issues/7297#issuecomment-1446041345
the aim here is to prevent *any* tests to, knowingly or unknowingly,
modify a user's global git-configuration.

The `check_gitconfig_global` runs on every test and compares pre/post
checksums of the global git-config source file. An assertion
error is raised for any detected change.

The aim here is to identify potential problems.