New `datalad_cfg` fixture
It provides a `ConfigManager` instance that uses a temporary global
Git configuration for the scope of a single test. This instance
is actually the global `ConfigManager` at `datalad.cfg` that is reloaded
twice (before and after a test) to make this temporary configuration
active for any configuration queries and manipulation of the global
manager during the test runtime.

An existing global configuration on disk is NOT modified. Instead
`GIT_CONFIG_GLOBAL` is set to point to a temporary location. This
approach minimized interference with Git operations outside the
test battery execution.

Because support for `GIT_CONFIG_GLOBAL` was only introduced with Git
2.32, and test using this fixture is skipped on older versions.

Ping https://github.com/datalad/datalad/issues/7297