Replace `memory_keyring` with `tmp_keyring`
The former was limited to within process operations. However, we do need
to be able to communicate credentials to child processes, such as
(special) remotes.

The approach for the `tmp_keyring` fixture is to replace the storage
file for the plaintext keyring with a random file, and then communicate
the path via the DATALAD_TESTS_TMP_KEYRING_PATH envvar.

An additional patch recognizes this variable and alters a keyring
in a child process accordingly.