Make `EnsureGeneratorFromFileLike` more suitable for batch-mode use
Previously, any invalid input caused an immediate and unrecoverable stop
and raise.

However for any batch mode application it is essential to be able to
ignore errors and/or simply report on them

This change adds an option to yield (not raise) any occuring internal
exception. This is straightforward to distinuish from regular items.
They can be reraised outside, or ignored, or reported on.

Closes #273