Detail `download()` input specification
Now also all aspects of `path` resolution, and deriving a filename
from an input URL are done outside the command implementation and
in respective constraints.

Improtantly, this now also uses a new feature of
`EnsureGeneratorFromFileLike` that can now simply relay per-item
exceptions to a caller (the `download()` command in this case).
This allows for performing comprehensive item constraint evaluation,
have that use regular exceptions (as if done upfront, but here actually
during the runtime of a command, due to the long-running generator
providing command input AKA batch mode). The command gets the exception
yielded (not raised) and can decide what to do with it. `download()`
maintains its previous behavior and reports an 'impossible' result.