Adjust/expand `download()` tests
The command can now immediately error when it is clear from the input
arguments that no destination filename can be derived (when none is
provided).

When this is not possible, the command continue to yield an 'impossible'
result.

RF tests on invalid command calls into a dedicated function. No need for
an http server in this case.