Rectify OR and AND of `Constraints` and `AltConstraints`
Both implementations performed in-place modifications of the
left-side constraint. This is not appropriate, because it matches
the semantics of `__iand__` and `__ior__`, rather than
`__and__` and `__or__`.

This change move away from the in-place implementation and simply
returns a new instance. This is cheap.

There is no implementation of `__iadd__` and `__ior__`, although
simply renaming the methods would have been enough.

There is no need for it, and it can be done whenever there is actual
demand.

Closes #291