New `WithDescription` meta/compound constraint
This is the main approach to tailor to self-description of a constraint
to a non-standard, unforseen, or simply unsupported use case.

It wraps any other constraint and provides the documentation it has been
parameterized with, instead of the original ones. A full replacement.

For tailoring wrapped constraints to a dataset-scope, a dedicated set
of documentation overrides with the tailored constraint can be given
in addition.

Closes #267