Demo custom constraint errors on `download` command
This command has one of the most complex input specification implemented
as nested constraints. Here is a demo of a (semi-structured) error message
rendering before this change:

```
❯ datalad download wurst
[ERROR  ] 1 command parameter constraint violation
spec=['wurst']
  does not match any of 3 alternatives
    - does not match any of 2 alternatives
      - not a recognized mapping
      - does not match any of 2 alternatives
        - the JSON object must be str, bytes or bytearray, not list
        - not a string
    - does not match any of 2 alternatives
      - not enough values to unpack (expected 2, got 1)
      - does not match any of 2 alternatives
        - Expecting value: line 1 column 1 (char 0)
        - URL is missing 'scheme' component
    - not '-', or a path to an existing file
```

There is a lot of information already, but clarity is lacking.

This change applied the `WithDescription` meta constraint to bring
clarity to the error messages. The new message for the same invalid
call used above now looks like this

```
❯ datalad download wurst
[ERROR  ] 1 command parameter constraint violation
spec=['wurst']
  does not provide URL->(PATH|-) mapping(s)
    - not a single item
      - not a dict, length-2-iterable, or space-delimited str
      - not a URL with a path component from which a filename can be derived
      - not a JSON-encoded str with an object or length-2-array
    - not a list of any such item
    - not a path to a file with one such item per-line, nor '-' to read any such item from STDIN
```

The custom message use the fact that the order of constraints and
therefore the order in which violations are detected is deterministic.
The respective individual message can reference prior messages such that
the result is a more cohesive whole that, importantly, can avoid
needless repetition of information.

Conceptually (and technically), although this message is composed of the
output of multiple constraints, it is neverthess still a *single*
message, produced by a *single* top-level constraint (`AnyOf`).
Therefore, we need not worry much about this information being taken
apart and thereby becoming harder to comprehend or incomplete.

The particular set of example message may inform #274 and guidelines on
how constraints and their messaging should be implemented.