Report full "address" of callable with deprecation
The previous shortening to a class name when the callable is a class
`__call__` makes sense when thinking about DataLad `Interface` classes.
However, in general it is not good (and this helper is generic).

The audience for deprecation messages are developers and developer
appreciate precision over riddles. Saying that the parameter
`p1` of `ClassName` is deprecated, makes it look as if
`ClassName.__init__` is the actual callable, but it would not be.

Given that a class name alone cannot be considered unqiue, the full
address (incl. module) is used in the message now.

Tests are adjusted accordingly.