Make version declaration mandatory as kwarg specification
I presently consider it better to make the nature of the version
declaration more obvious in the resulting code. It is already mandatory
(which is essential, I think).

A custom message is also useful to be mandatory. Deprecation without
guidance on adjustment is bad practice.

Given that there are two essential items, let's have the different
semantics be obvious in each declaration.