Make special remote processes report progress logs to git-annex
Add dedicated log handler is attached to both the datalad "root" logger
and the special remote class instance. The handler gets a filter to only
makes it see progress log records (this is something that also the main
progress handler should be doing). In the generic main() of special
remotes, this log setup modifications is added as an additional log
output -- no interference with any existing setup.

This sketch is limited/incomplete in that it assumes only a single
progress tracker to be used (although it is already set up to support
more than one). So in case of multiple concurrent trackers, it would
send rubbish progress logs to git-annex.

That being said, with this change we see git-annex native progress
reporting with git annex get and datalad native progress reporting with
datalad get.

Closes #328