TMP: Document a note to not set version_aware fs_kwargs
When working with S3 URLs, s3fs employs internal URL parsing to detect
whether the S3 file has versioning enabled. Based on the outcome of this
evaluation, it sets the version_aware value itself.
If a user sets this keys' value to False in fs_kwargs of the URL handler,
but then supplies a versioned URL, s3fs's attempt to set this key to True
is sabotaged, and the download crashes:

datalad_next.url_operations.UrlOperationsRemoteError: UrlOperationsRemoteError for 's3://mslw-datalad-test0-versioned/3versions-allversioned.txt?versionId=Tro_UjqVFJfr32v5tuPfjwtOzeqYCxi2'

To leave a trace of this, even if only temporary, this commit adds a short
paragraph to the URL handler's docstring to warn about setting this key.